Sửa lại một vài chỗ trong framework và làm thêm vài animation cho Samus tạo thêm class World và class GroupObject, Collider Xong va chạm

+++ METROID-master/WinMain/Manager.cpp
@@ -15,12 +15,12 @@ void Manager::CheckDistance()
 
 }
 
-void Manager::_CreateBullets(int X, int Y, int VX, int VY)
+void Manager::_CreateBullets(float X, float Y, float VX, float VY)
 {
 	weapon.push_back( new Bullet( &m_d3ddv, X, Y, VX, VY) );
 }
 
-void Manager::UpdateObject(int delta)
+void Manager::UpdateObject(float delta)
 {
 	for (int i = 0; i < weapon.size(); i++)
 	{
